Add files via upload
A perfect ToDoList is where one can add tasks, edit tasks and delete tasks with good hovering effects and a user friendly interactive page. This project full fills all the requirements mentioned. It has the ability to store the tasks in local storage so that user could get back their data though the page is refreshed.